backend/requirements.txt
uuid==1.30
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
astroid==3.0.0
attrs==23.2.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.8
convertapi==1.8.0
distlib==0.3.7
Django==4.2.5
fastapi==0.116.1
filelock==3.12.4
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
graphviz==0.20.1
h11==0.16.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
multidict==6.0.5
numpy==1.26.3
pipenv==2023.9.8
platformdirs==3.10.0
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.5.1
PyJWT==2.8.0
pyobjc==9.2
pyobjc-core==9.2
pyobjc-framework-Accessibility==9.2
pyobjc-framework-Accounts==9.2
pyobjc-framework-AddressBook==9.2
pyobjc-framework-AdServices==9.2
pyobjc-framework-AdSupport==9.2
pyobjc-framework-AppleScriptKit==9.2
pyobjc-framework-AppleScriptObjC==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-AppTrackingTransparency==9.2
pyobjc-framework-AudioVideoBridging==9.2
pyobjc-framework-AuthenticationServices==9.2
pyobjc-framework-AutomaticAssessmentConfiguration==9.2
pyobjc-framework-Automator==9.2
pyobjc-framework-AVFoundation==9.2
pyobjc-framework-AVKit==9.2
pyobjc-framework-AVRouting==9.2
pyobjc-framework-BackgroundAssets==9.2
pyobjc-framework-BusinessChat==9.2
pyobjc-framework-CalendarStore==9.2
pyobjc-framework-CallKit==9.2
pyobjc-framework-CFNetwork==9.2
pyobjc-framework-ClassKit==9.2
pyobjc-framework-CloudKit==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Collaboration==9.2
pyobjc-framework-ColorSync==9.2
pyobjc-framework-Contacts==9.2
pyobjc-framework-ContactsUI==9.2
pyobjc-framework-CoreAudio==9.2
pyobjc-framework-CoreAudioKit==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-CoreData==9.2
pyobjc-framework-CoreHaptics==9.2
pyobjc-framework-CoreLocation==9.2
pyobjc-framework-CoreMedia==9.2
pyobjc-framework-CoreMediaIO==9.2
pyobjc-framework-CoreMIDI==9.2
pyobjc-framework-CoreML==9.2
pyobjc-framework-CoreMotion==9.2
pyobjc-framework-CoreServices==9.2
pyobjc-framework-CoreSpotlight==9.2
pyobjc-framework-CoreText==9.2
pyobjc-framework-CoreWLAN==9.2
pyobjc-framework-CryptoTokenKit==9.2
pyobjc-framework-DataDetection==9.2
pyobjc-framework-DeviceCheck==9.2
pyobjc-framework-DictionaryServices==9.2
pyobjc-framework-DiscRecording==9.2
pyobjc-framework-DiscRecordingUI==9.2
pyobjc-framework-DiskArbitration==9.2
pyobjc-framework-DVDPlayback==9.2
pyobjc-framework-EventKit==9.2
pyobjc-framework-ExceptionHandling==9.2
pyobjc-framework-ExecutionPolicy==9.2
pyobjc-framework-ExtensionKit==9.2
pyobjc-framework-ExternalAccessory==9.2
pyobjc-framework-FileProvider==9.2
pyobjc-framework-FileProviderUI==9.2
pyobjc-framework-FinderSync==9.2
pyobjc-framework-FSEvents==9.2
pyobjc-framework-GameCenter==9.2
pyobjc-framework-GameController==9.2
pyobjc-framework-GameKit==9.2
pyobjc-framework-GameplayKit==9.2
pyobjc-framework-HealthKit==9.2
pyobjc-framework-ImageCaptureCore==9.2
pyobjc-framework-IMServicePlugIn==9.2
pyobjc-framework-InputMethodKit==9.2
pyobjc-framework-InstallerPlugins==9.2
pyobjc-framework-InstantMessage==9.2
pyobjc-framework-Intents==9.2
pyobjc-framework-IntentsUI==9.2
pyobjc-framework-IOBluetooth==9.2
pyobjc-framework-IOBluetoothUI==9.2
pyobjc-framework-IOSurface==9.2
pyobjc-framework-iTunesLibrary==9.2
pyobjc-framework-KernelManagement==9.2
pyobjc-framework-LatentSemanticMapping==9.2
pyobjc-framework-LaunchServices==9.2
pyobjc-framework-libdispatch==9.2
pyobjc-framework-libxpc==9.2
pyobjc-framework-LinkPresentation==9.2
pyobjc-framework-LocalAuthentication==9.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==9.2
pyobjc-framework-MailKit==9.2
pyobjc-framework-MapKit==9.2
pyobjc-framework-MediaAccessibility==9.2
pyobjc-framework-MediaLibrary==9.2
pyobjc-framework-MediaPlayer==9.2
pyobjc-framework-MediaToolbox==9.2
pyobjc-framework-Metal==9.2
pyobjc-framework-MetalFX==9.2
pyobjc-framework-MetalKit==9.2
pyobjc-framework-MetalPerformanceShaders==9.2
pyobjc-framework-MetalPerformanceShadersGraph==9.2
pyobjc-framework-MetricKit==9.2
pyobjc-framework-MLCompute==9.2
pyobjc-framework-ModelIO==9.2
pyobjc-framework-MultipeerConnectivity==9.2
pyobjc-framework-NaturalLanguage==9.2
pyobjc-framework-NetFS==9.2
pyobjc-framework-Network==9.2
pyobjc-framework-NetworkExtension==9.2
pyobjc-framework-NotificationCenter==9.2
pyobjc-framework-OpenDirectory==9.2
pyobjc-framework-OSAKit==9.2
pyobjc-framework-OSLog==9.2
pyobjc-framework-PassKit==9.2
pyobjc-framework-PencilKit==9.2
pyobjc-framework-PHASE==9.2
pyobjc-framework-Photos==9.2
pyobjc-framework-PhotosUI==9.2
pyobjc-framework-PreferencePanes==9.2
pyobjc-framework-PushKit==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-QuickLookThumbnailing==9.2
pyobjc-framework-ReplayKit==9.2
pyobjc-framework-SafariServices==9.2
pyobjc-framework-SafetyKit==9.2
pyobjc-framework-SceneKit==9.2
pyobjc-framework-ScreenCaptureKit==9.2
pyobjc-framework-ScreenSaver==9.2
pyobjc-framework-ScreenTime==9.2
pyobjc-framework-ScriptingBridge==9.2
pyobjc-framework-SearchKit==9.2
pyobjc-framework-Security==9.2
pyobjc-framework-SecurityFoundation==9.2
pyobjc-framework-SecurityInterface==9.2
pyobjc-framework-ServiceManagement==9.2
pyobjc-framework-SharedWithYou==9.2
pyobjc-framework-SharedWithYouCore==9.2
pyobjc-framework-ShazamKit==9.2
pyobjc-framework-Social==9.2
pyobjc-framework-SoundAnalysis==9.2
pyobjc-framework-Speech==9.2
pyobjc-framework-SpriteKit==9.2
pyobjc-framework-StoreKit==9.2
pyobjc-framework-SyncServices==9.2
pyobjc-framework-SystemConfiguration==9.2
pyobjc-framework-SystemExtensions==9.2
pyobjc-framework-ThreadNetwork==9.2
pyobjc-framework-UniformTypeIdentifiers==9.2
pyobjc-framework-UserNotifications==9.2
pyobjc-framework-UserNotificationsUI==9.2
pyobjc-framework-VideoSubscriberAccount==9.2
pyobjc-framework-VideoToolbox==9.2
pyobjc-framework-Virtualization==9.2
pyobjc-framework-Vision==9.2
pyobjc-framework-WebKit==9.2
python-dotenv==1.1.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
sqlparse==0.4.4
starlette==0.47.1
twilio==9.2.3
typing-inspection==0.4.1
typing_extensions==4.13.0
urllib3==2.0.5
uvicorn==0.35.0
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
Werkzeug==3.1.3
yarl==1.9.4
selenium==4.26.1
webdriver-manager==4.0.2
wakeonlan==1.1.6
